requirements.txt
requests==2.32.3
tenacity==9.0.0
polars==1.26.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
pendulum==3.1.0
pyyaml==6.0.2
apache-airflow==2.10.5
python-dotenv==1.0.1